Split snapshot workload package by resource type
